Create StudentDTO.java
The StudentDTO class is a Data Transfer Object (DTO) used to encapsulate data related to a student, such as their ID, student number, GPA, and name. This class is necessary for transferring student data between different layers of an application, such as between the data access layer ,is kind of constructor base file.