Update Main.java
Implemented the functionality to update a student's information by ID in the Main class. Added a case 5 in the menu to handle updating student details. Prompted the user to enter the student ID to update and the new student number, GPA, and name. Created a new StudentDTO object with the updated information and called the updateStudentById method from StudentDAOImpl to perform the update in the database