Create Client.java
Created a Client file and added functionality to the Client class to interact with the server. The client displays a menu allowing the user to choose between displaying an entity by ID, displaying all entities, or exiting the program. Depending on the user's choice, the client sends the corresponding request to the server and receives JSON data in response. The displayEntityById and displayAllEntities methods handle the communication with the server for each option. This implementation enables users to interact with the server and retrieve data based on their preferences.