Update Server.java
Implemented the sendImagesList method to send a list of image file names available for download from the server. If the images directory exists and is a directory, it retrieves the list of image file names, converts it to JSON format using JsonConverter.arrayToJson, and sends it to the client. If the directory does not exist or is not a directory, it sends an empty JSON array ("[]"). And also added the request type 4 to return sendImagesList to the client page.