Update Client.java
mplemented functionality to retrieve a list of image file names from the server and download a selected image file. The getImagesList method sends a request to the server, receives JSON data containing the list of image file names, prompts the user to enter the name of the image file to download, and then sends a request to the server with the selected file name. If the file exists on the server, its size and data are received, and the file is saved locally. Otherwise, a message indicating that the file was not found is displayed.